# refine svmpredict