Ref:  https://github.com/JuliaML/LIBSVM.jl/pull/76#discussion_r650039472

~I imported the function `libsvm_predict_probability` and `libsvm_predict_values` for temporary and discussion needs.
They should be relocated later in #76.~ Rebased.