@barucden 2021-06-11
It would be great if we managed to make `svmpredict` more type-stable. Currently, the type of `pred` is uncertain, as well as `output` (before your commit). Perhaps we could define some auxiliary functions as they do [here](https://docs.julialang.org/en/v1/manual/performance-tips/#kernel-functions)? I believe it would result not only in more performant but also cleaner code.