@iblis17 2021-06-12
> It would be great if we managed to make `svmpredict` more type-stable.

I added `svmpredict_fill!`, it became slightly faster.

Before:
```julia
julia> A = rand(2, 2_000_000);

julia> @benchmark svmpredict($model, $A)
BenchmarkTools.Trial:
  memory estimate:  152.59 MiB
  allocs estimate:  15
  --------------
  minimum time:     1.185 s (0.10% GC)
  median time:      1.207 s (1.83% GC)
  mean time:        1.236 s (3.57% GC)
  maximum time:     1.305 s (7.68% GC)
  --------------
  samples:          5
  evals/sample:     1
```

After:
```julia
julia> @benchmark svmpredict($model, $A)
BenchmarkTools.Trial:
  memory estimate:  152.59 MiB
  allocs estimate:  16
  --------------
  minimum time:     1.062 s (0.11% GC)
  median time:      1.092 s (2.34% GC)
  mean time:        1.113 s (4.13% GC)
  maximum time:     1.195 s (8.72% GC)
  --------------
  samples:          5
  evals/sample:     1
```