@iblis17 2021-06-12
Given that there are some CI failures popped out randomly, I suspect that there are some GC safety issues

e.g.
https://github.com/JuliaML/LIBSVM.jl/pull/77/checks?check_run_id=2807830407