@iblis17 2021-06-12
I cannot reproduce that on my machine either.

I can add some `GC.@preserve` in other PR and let's see the random error pop out in future or not.